복잡성을 줄이기 위해 외부로 분리했던 GameMap 생성 전략을 LadderGame 클래스로 이동
LadderGameMapStretegy, DefaultGameMapStretegy 의 기능을 LadderGame 으로 이동하였습니다.
