--- src/main/java/app/jinan159/ladder/stretegy/LadderGameMapStretegy.java
@@ -1,9 +1,0 @@
-package app.jinan159.ladder.stretegy;
-
-import app.jinan159.ladder.meta.LadderElement;
-
-public interface LadderGameMapStretegy {
-
-        LadderElement getLadderElementOnPotision(int x, int y);
-
-}
